LayoutTests/accessibility/transformed-bounds-expected.txt

Test
This tests that if a view is in a transformed layer, the bounds will not be offset by the origin of the transformed layer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tab.width < 40 is true
PASS tab.height < 30 is true
PASS successfullyParsed is true

TEST COMPLETE
